+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 |Release Version|Date|Description
 |---------------|----|-----------|
+|1.1.5|2013/05/01|Added uptime option|
+|1.1.4|2013/05/01|Added update option|
+|1.1.3|2013/05/01|Added check option|
 |1.1.2|2013/03/08|Added stat/timesync_status/version_old_installer|
 |1.1.2|2013/03/07|Cchanged useage display/added remove_old_installer option/rename installer_version to version_installer|
 |1.1.1|2013/03/06|VMware Tools version independent check & better user input checking|